--- a/crystal/raze/src/server.cr
+++ b/crystal/raze/src/server.cr
@@ -15,4 +15,5 @@ end
 Raze.config.logging = false
 Raze.config.port = 3000
 Raze.config.env = "production"
+Raze.config.reuse_port = true
 Raze.run