Hi,

This `PR` aims to implements (or check) whole CPU usage, it closes #69

This feature **COULD** be :
+ `SO_REUSEPORT` via `spawn`/`fork`
+ a custom implementation
+ ...

+ [x] nim
   + [x] jester
+ [x] elixir
   + [x] plug
   + [x] phoenix
+ [x] go
   + [x] fasthttprouter
   + [x] violetear
   + [x] kami
   + [x] gorilla-mux
   + [x] gin
   + [x] chi
   + [x] beego
   + [x] gf
   + [x] echo
+ [ ] kotlin
   + [ ] ktor
+ [ ] swift
   + [ ] kitura
   + [ ] perfect
   + [ ] vapor
+ [x] php
  + [x] symfony
  + [x] zend-framework
  + [x] laravel
  + [x] lumen
  + [x] zend-expressive
  + [x] slim
+ [ ] scala
   + [x] akkahttp
   + [x] http4s
+ [ x] rust
   + [ ] iron
   + [x] gotham
   + [ ] rocket
   + [ ] actix-web
   + [ ] nickel
+ [ ] objc
   + [ ] criollo
+ [x] python
   + [x] aiohttp
   + [x] tornado
   + [x] flask
   + [x] django
   + [x] vibora
   + [x] molten
   + [x] hug
   + [x] fastapi
   + [x] quart
   + [x] responder
   + [x] cyclone
   + [x] bocadillo
   + [x] sanic
   + [x] bottle
   + [x] falcon
   + [x] japronto
   + [x] starlette
+ [] crystal
   + [x] onyx
   + [x] orion
   + [ ] athena
   + [x] kemal
   + [x] router.cr
   + [x] spider-gazelle
   + [x] lucky
   + [x] amber
+ [ ] csharp
   + [ ] aspnetcore
+ [x] node
   + [x] polka
   + [x] restana
   + [x] turbo_polka
   + [x] restify
   + [x] foxify
   + [x] express
   + [x] koa
   + [x] rayo
   + [x] muneem
   + [x] fastify
   + [x] hapi
+ [x] java
   + [x] act
+ [x] cpp
   + [x] evhtp
+ [x] ruby
   + [x] hanami
   + [x] rails
   + [x] flame
   + [x] agoo
   + [x] roda
   + [x] sinatra
   + [x] rack-routing
   + [x] cuba
+ [ ] c
  + [x] agoo-c
  + [ ] kore


Regards,