@waghanza 2019-01-24
I'm adding `SO_REUSEPORT` for ruby's **framework**
https://github.com/puma/puma/pull/1712