@dom96 2019-01-24
Jester through httpbeast uses SO_REUSEPORT implicitly as long as threads are enabled via ``--threads:on``.