@ohler55 2019-01-24
Agoo-C uses `SO_REUSEPORT` as well (`src/agoo/bind.c:236`)