@kataras 2019-01-25
Not by-default, but they can, examples:

- For Iris: https://github.com/kataras/iris/blob/master/_examples/http-listening/custom-listener/unix-reuseport/main.go
- For Fasthttp: https://github.com/valyala/fasthttp/blob/master/reuseport/reuseport_example_test.go