@johngcn 2019-01-25
GF's using `net/http` as its underlying HTTP server currently, and it's a pity that `net/http` just supports `SO_REUSEADDR` instead of `SO_REUSEPORT`.