@waghanza 2019-01-25
@johncn I thought `SO_REUSEPORT` was included in **stdlib**

https://go-review.googlesource.com/c/go/+/37039/

Thanks @kataras for the tip. I think it's better to compare framework here having this feature enable in all impl.