@johngcn 2019-01-26
@waghanza 

Thank you for the tips. I see, there're some workarounds for implementing `SO_REUSEPORT` in golang using `syscall`, in some ARCHs. 

I know it's efficient for application using multi-process or multi-thread when its `SO_REUSEPORT` option enabled. But I doubt the performance in single process when such option enabled in golang, as there're just goroutines, you cannot control the underlying threads. 

It may be quite efficient  using `SO_REUSEPORT` for those golang applications, which are designed and running in processes. How can you limit the resources when running the performance comparison?