@ioquatix 2019-01-28
@waghanza It does now! Use v0.22.0.

`falcon serve --reuse-port` - it's not the default for obvious reasons.