@ioquatix 2019-01-28
BTW, do you mind explaining what this is for?

I actually used to use SO_REUSEPORT but I found that on Darwin it doesn't work as expected - the OS always uses the first process to bind to the port (but it's not an error to start other processes). The better model is to bind a socket and share it over multiple threads/processes.