@waghanza 2019-01-28
> What are you trying to equalise?

I think it's better to have `SO_REUSEPORT` everywhere, or no where here ;-)

I'm trying to maximise efficiency (use the maximum of server capacity)