@ioquatix 2019-01-28
So, I agree with your basic idea: make everything equal.

But, unless I'm missing something important, I don't see how `SO_REUSEPORT` achieves this.

Using `--reuse-port` with falcon won't change performance I the slightest.

Whether you use that or not, it only binds one port, and then shares it over N processes/threads (N = processor core count).

The only difference is you can start multiple falcon processes bound to the same port, which won't change performance but would make it confusion as to which process will serve which request.

Can you explain what you think using `SO_REUSEPORT` achieves? Are you planning to start N processes for single-process servers?