@waghanza 2019-01-28
@ioquatix Using `SO_REUSEPORT` can make a performance increase (and probably reduce resources usage)
https://github.com/puma/puma/pull/1712

As far a I understand, `SO_REUSEPORT` make several process use the same port, so the kernel has not to make some syscall to attach a port to a process, so less syscalls, so less resources used, but @OvermindDL1 will have a better explanation that mine