@ioquatix 2019-01-28
@waghanza what you've said makes no sense at all. Sorry, it simply doesn't align up with what SO_REUSEPORT does.

If anything, `SO_REUSEPORT` is a crappy way to achieve multi-process or multi-threaded servers. Maybe it's useful for rolling restarts. But it has nothing to do with improving performance of a well designed server (e.g. bind before fork or bind before threads).

Here is the evidence from my testing:

Without `--reuse-port`:

```
Running falcon with 128 concurrent connections...
Running 2s test @ http://127.0.0.1:9292/small
  8 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.66ms    2.90ms  38.84ms   92.00%
    Req/Sec     7.59k     2.56k   35.61k    95.65%
  121555 requests in 2.08s, 143.51MB read
Requests/sec:  58520.43
Transfer/sec:     69.09MB
```

With `--reuse-port`:

```
Running falcon with 128 concurrent connections...
Running 2s test @ http://127.0.0.1:9292/small
  8 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.53ms    2.40ms  41.16ms   93.89%
    Req/Sec     7.37k     1.38k   21.34k    93.90%
  120275 requests in 2.10s, 142.00MB read
Requests/sec:  57291.82
Transfer/sec:     67.64MB
```