@waghanza 2019-01-28
@ioquatix I can see a slight performance increase it `req/sec`. For `latency`, I'm not sure `wrk` is in fact the right tool ;-) 
see https://github.com/the-benchmarker/web-frameworks/issues/670