@ioquatix 2019-01-28
The req/s actually dropped slightly and it's well within margin of error which would I'd say on any given run would be within 10%.

I agree latency computation is tricky. `wrk` isn't too bad.