@waghanza 2019-01-28
@ioquatix You get a point, you can suggest a tool to replace  `wrk` if you know one :stuck_out_tongue_winking_eye: