@ioquatix 2019-01-28
I reran my tests under some different benchmarks (big response, small response, tiny response). There was no obvious difference. That being said, if difference is small, it might not be obvious. Honestly, most web server that's capable of more than 50k req/s is fast enough in production. Franky speaking, from experience, as soon as you have any kind of database, it won't matter if you can serve 10k or 100k req/s, the overhead of the application server is so much bigger, the web server overhead becomes irrelevant.