@ioquatix 2019-01-28
Here is a benchmark tool I wrote, but it's limited a bit by the performance of Ruby. It's interesting though because it tries to find the limits of concurrency for a server within a % tolerance of performance degradation: https://github.com/socketry/benchmark-http

The way it works is it makes 1 request, and measures performance/latency.

Then it makes 2 request at the same time.

It does binary search to find the point at which your latency is impacted by a certain %. This is the point at which your server is slowing down multiple requests due to internal contention. It's an interesting metric because it even applies to slow servers over the internet - it's not about absolute latency but how latency is impacted as you increase concurrency.