@ioquatix 2019-01-28
That tool looks awesome. That being said, in order for a benchmark tool to be accurate, it must be coded in a low level language IMHO. That's why I don't trust `benchmark-http` for raw performance metrics. For other things (like concurrency), it's fine.