@ohler55 2019-01-28
There is also `perfer`. I wrote it to handle high performance C servers. https://github.com/ohler55/perfer