@proyb6 2019-01-28
I think this article may be worth reading on various benchmark tools.
https://blog.loadimpact.com/open-source-load-testing-tool-benchmarks-v2